Fixes #65: Add custom endpoint to get Online Applications Campaign
Remove entrypoint code and add tests

Update Application Form to use REST endpoint

Update Exception type

Remove closing tag
