--- package/src/custom/Extension/application/Ext/EntryPointRegistry/getCampaign.php
@@ -1,6 +1,0 @@
-<?php
-// Copyright 2018 SugarCRM Inc.  Licensed by SugarCRM under the Apache 2.0 license.
-$entry_point_registry['getOnlineApplicationsCampaign'] = array(
-    'file' => 'custom/src/application_campaign/GetCampaignEntrypoint.php',
-    'auth' => false
-);
