+++ package/src/custom/modules/Campaigns/clients/base/api/help/GetOnlineApplicationsCampaignId.html
@@ -1,0 +1,35 @@
+<h2>Overview</h2>
+<span class="lead">
+    A custom GET endpoint. Returns the ID of the most recently updated Campaign with the name 'Online Applications'.
+</span>
+
+<h2>Result</h2>
+<table class="table table-hover">
+    <thead>
+    <tr>
+        <th>Name</th>
+        <th>Type</th>
+        <th>Description</th>
+    </tr>
+    </thead>
+    <tbody>
+    <tr>
+        <td>
+            id
+        </td>
+        <td>
+            String
+        </td>
+        <td>
+            The ID of the most recently updated Campaign with the name 'Online Applications'.
+        </td>
+    </tr>
+    </tbody>
+</table>
+
+<h3>Output Example</h3>
+<pre class="pre-scrollable">
+{
+    "id": "f56b6d18-52e1-11e8-a080-34363bc46900"
+}
+</pre>
